pyproject.toml
[project]
name = "zenvox"
version = "1.1.0"
description = "Voice to text, cleaned by AI. Fast, private, always ready."
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "ZenSystem" },
]

# Core dependencies — always required
dependencies = [
    "customtkinter>=5.2",
    "numpy>=2.0",
    "pyautogui>=0.9",
    "pyperclip>=1.9",
    "pystray>=0.19",
    "sounddevice>=0.5",
    "Pillow>=12.0",
    "faster-whisper>=1.1",
    "keyring>=24.0",
    "pynput>=1.7",
]

[project.optional-dependencies]
# Cleaning providers — install the ones you want to use
gemini = ["google-genai>=1.0"]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.80"]
groq = ["groq>=0.9"]
# All providers
all = [
    "google-genai>=1.0",
    "openai>=1.0",
    "anthropic>=0.80",
    "groq>=0.9",
]
# GPU acceleration (NVIDIA CUDA) — includes required cuBLAS/cuDNN runtime DLLs on Windows
gpu = [
    "ctranslate2>=4.5",
    "nvidia-cublas-cu12",
    "nvidia-cudnn-cu12",
    "nvidia-cuda-runtime-cu12",
]

[project.scripts]
zenvox = "zenvox:main"

[tool.setuptools]
py-modules = ["zenvox", "config", "providers", "history"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
